+++ server/server.cpp
@@ -30,7 +30,7 @@ Server::Server(QString servername, QObject *parent) : QObject(parent){
             in.setVersion(QDataStream::Qt_5_0);
 
             if(sock->property("BTR").toInt() == 0){
-                quint16 btr;
+                quint64 btr;
                 in >> btr;
                 sock->setProperty("BTR", QVariant(btr));
             }
@@ -69,10 +69,10 @@ void Server::sendMessageToClient(int clientID, QString message){
 
     out.setVersion(QDataStream::Qt_5_0);
 
-    out << (quint16)0;
+    out << (quint64)0;
     out << message;
     out.device()->seek(0);
-    out << (quint16)(block.size() - sizeof(quint16));
+    out << (quint64)(block.size() - sizeof(quint64));
 
     qint64 c = sock->write(block);
     sock->waitForBytesWritten();
